# backend/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'VanScape API'
  description: 'Vanscape API endpoints'
  version: 1.0.0
servers:
  -
    url: 'https://api-vanscape.duckdns.org'
    description: 'API over HTTPS'
paths:
  '/api/friendships/{friendId}':
    post:
      tags:
        - Friendship
      summary: 'Creates a new friendship between the authenticated user and the user with the given friendId.'
      operationId: post_create_friendship
      parameters:
        -
          name: friendId
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9]+'
      responses:
        '201':
          description: 'Returns data for the created friendship'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendshipDTO'
        '400':
          description: 'Errors found in received data in %s'
        '401':
          description: 'No authenticated user found in JWT'
        '404':
          description: 'Friendship not found'
        '409':
          description: 'Friendship already exists'
        '500':
          description: 'Oops ! Something went wrong.'
    delete:
      tags:
        - Friendship
      summary: 'Deletes a friendship between the authenticated user and the specified friend.'
      operationId: delete_delete_friendship
      parameters:
        -
          name: friendId
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9]+'
      responses:
        '204':
          description: 'Successfully deleted the friendship'
        '401':
          description: 'No authenticated user found in JWT'
        '404':
          description: 'Friendship not found'
        '500':
          description: 'Oops ! Something went wrong.'
  '/api/friendships/pending/{type}':
    get:
      tags:
        - Friendship
      summary: 'Retrieves a list of pending friendship requests (received or sent) for the authenticated user.'
      operationId: get_read_pending_friendships
      parameters:
        -
          name: type
          in: path
          required: true
          schema:
            type: string
            pattern: received|sent
            enum:
              - received
              - sent
      responses:
        '200':
          description: 'Returns a list of pending friendship requests (received or sent)'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PartialFriendshipDTO'
        '401':
          description: 'No authenticated user found in JWT'
        '500':
          description: 'Oops ! Something went wrong.'
  /api/friendships/pending/received/summary:
    get:
      tags:
        - Friendship
      summary: 'Retrieves the number of pending friendship requests received by the authenticated user.'
      operationId: get_read_pending_friendships_summary
      responses:
        '200':
          description: 'Returns the number of pending friendship requests received'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendshipReceivedSummaryDTO'
        '401':
          description: 'No authenticated user found in JWT'
        '500':
          description: 'Oops ! Something went wrong.'
  /api/friendships/confirmed:
    get:
      tags:
        - Friendship
      summary: 'Retrieves the list of confirmed friendships for the authenticated user.'
      operationId: get_read_confirmed_friendships
      responses:
        '200':
          description: 'Returns a list of confirmed friendship'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PartialFriendshipDTO'
        '401':
          description: 'No authenticated user found in JWT'
        '500':
          description: 'Oops ! Something went wrong.'
  '/api/friendships/{requesterId}/confirm':
    patch:
      tags:
        - Friendship
      summary: 'Confirms a friendship request by setting its `isConfirmed` status to true.'
      operationId: patch_update_confirm_friendship
      parameters:
        -
          name: requesterId
          in: path
          required: true
          schema:
            type: string
            pattern: '[0-9]+'
      responses:
        '200':
          description: 'Return data for the updated friendship'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendshipDTO'
        '401':
          description: 'No authenticated user found in JWT'
        '404':
          description: 'Friendship not found'
        '500':
          description: 'Oops ! Something went wrong.'
...
components:
  schemas:
...
    FriendshipDTO:
      title: Friendship
      description: 'Details of a friendship'
      required:
        - requester
        - receiver
        - isConfirmed
      properties:
        requester:
          description: 'Requester identity'
          oneOf:
            -
              $ref: '#/components/schemas/FriendshipUserDTO'
        receiver:
          $ref: '#/components/schemas/FriendshipUserDTO'
        isConfirmed:
          description: 'Friendship is confirmed or not'
          type: boolean
      type: object
    PartialFriendshipDTO:
      title: 'Partial Friendship'
      description: 'Details of a friendship without details of requester/receiver'
      required:
        - friend
        - isConfirmed
      properties:
        friend:
          $ref: '#/components/schemas/FriendshipUserDTO'
        isConfirmed:
          description: 'Friendship is confirmed or not'
          type: boolean
      type: object
    FriendshipReceivedSummaryDTO:
      title: 'Friendship received summary'
      description: 'Number of received friendship request'
      required:
        - count
      properties:
        count:
          description: 'Number of received friendship request'
          type: integer
          example: 3
      type: object
...
    FriendshipUserDTO:
      title: 'Friendship user identity'
      description: 'Identity of a friend'
      required:
        - id
        - pseudo
      properties:
        id:
          description: 'User id'
          type: integer
          example: 2
        pseudo:
          description: 'User pseudo'
          type: string
          example: 'Jane Doe'
        picture:
          description: 'Choosen avatar for profil picture'
          type: string
          nullable: true
      type: object
...
  securitySchemes:
    Bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer
security:
  -
    Bearer: []
...